frontの不要ファイルと行を削除+ tailwindのフォーマットツール導入
