+++ frontend/src/App.tsx
@@ -1,19 +1,15 @@
-import React from 'react';
+import React, { VFC } from 'react';
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { ReactQueryDevtools } from 'react-query/devtools';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import './App.css';
-import { VFC,useEffect} from 'react';
-import { BrowserRouter, Route, Switch} from 'react-router-dom'
-import { QueryClient, QueryClientProvider } from 'react-query'
-import { ReactQueryDevtools } from 'react-query/devtools'
-
-import Header from './components/Header'
+import Header from './components/Header';
+import PostForm from './components/posts/PostCreate';
 import PostShow from './components/posts/PostShow';
-import PostForm from './components/posts/PostForm';
-import Test from './components/test'
+import PostsIndex from './components/posts/PostsIndex';
 import ProfileEdit from './components/profile/ProfileEdit';
 import ProfileShow from './components/profile/ProfileShow';
-import PostsIndex from './components/posts/PostsIndex';
-import { useQueryPosts } from './hooks/reactQuery/useQueryPosts';
-
+import Test from './components/test';
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -27,10 +23,10 @@ const queryClient = new QueryClient({
 const App: VFC = () => {
   return (
     
-    <div className='relative h-screen'>
+  <div className='relative h-screen'>
     <QueryClientProvider client={queryClient}>
-    <BrowserRouter>
-    <Header/>
+      <BrowserRouter>
+        <Header/>
         <Switch>
           <Route exact path="/posts">
             {/* <MainTag/> */}
@@ -46,15 +42,7 @@ const App: VFC = () => {
           match={match}
           />
           }
-          />
-          <Route
-          exact path="/posts/edit/:postId"
-          render={({ match }:any) =>
-          <PostForm
-          match={match}
-          />
-          }
-          />
+          /> 
           <Route exact path="/profile/edit/:profileId"
           render={({ match }:any) =>
           <ProfileEdit
@@ -73,11 +61,10 @@ const App: VFC = () => {
             <Test/>
           </Route>
         </Switch>
-        </BrowserRouter>
-    <ReactQueryDevtools initialIsOpen={false} />
+      </BrowserRouter>
+      <ReactQueryDevtools initialIsOpen={false} />  
     </QueryClientProvider>
-    </div>
-
+  </div>
   );
 }
