+++ frontend/src/components/posts/PostCreateItem/FormRadio.tsx
@@ -1,12 +1,11 @@
-import React from 'react';
+import FormControl from '@material-ui/core/FormControl';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import FormLabel from '@material-ui/core/FormLabel';
 import Radio from '@material-ui/core/Radio';
 import RadioGroup from '@material-ui/core/RadioGroup';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import FormControl from '@material-ui/core/FormControl';
-import FormLabel from '@material-ui/core/FormLabel';
-import { useAppSelector, useAppDispatch } from '../../../app/hooks'
-import { setEditedPost, selectPost } from '../../../slices/postSlice'
-import { useMutatePost } from '../../../hooks/castomHook/useMutatePost'
+import React from 'react';
+import { useAppDispatch, useAppSelector } from '../../../app/hooks';
+import { selectPost, setEditedPost } from '../../../slices/postSlice';
 
 export const FormRadio = () => {
   const editedPost = useAppSelector(selectPost)
