+++ frontend/src/components/posts/PostIndexItem/AutoCompGenre.tsx
@@ -1,12 +1,11 @@
-import React from 'react';
 import TextField from '@material-ui/core/TextField';
 import Autocomplete from '@material-ui/lab/Autocomplete';
-import { useAppDispatch,useAppSelector} from '../../../app/hooks';
-import { resetQueryGenre, selectQueryGenre, setQueryGenre } from '../../../slices/postSlice';
+import React from 'react';
+import { useAppDispatch } from '../../../app/hooks';
+import { resetQueryGenre, setQueryGenre } from '../../../slices/postSlice';
 
 export const AutoCompGenre = () => {
   const dispatch = useAppDispatch()
-  const genre = useAppSelector(selectQueryGenre)
   const handleChange = (e:any,v:any) =>{
     v ?
     dispatch(setQueryGenre(v.genre))
