+++ frontend/src/components/profile/ProfileShow.tsx
@@ -1,13 +1,12 @@
 import React from 'react'
-import ProfileShowProfile from './ProfileShowItem/ProfileShowProfile'
-import ProfileShowPosts from './ProfileShowItem/ProfileShowPosts'
 import { useQueryFollows } from '../../hooks/reactQuery/useQueryFollows'
+import ProfileShowPosts from './ProfileShowItem/ProfileShowPosts'
+import ProfileShowProfile from './ProfileShowItem/ProfileShowProfile'
 
 const ProfileShow = ({match}:any) => {
-  // console.log(data)
   const id = match.params.profileId
   useQueryFollows(id)
-
+  
   return (  
     <div className='w-full py-2 lg:mx-auto'>
       <ProfileShowProfile id={id}/>
