+++ frontend/src/components/profile/ProfileShowItem/ProfileTabs.tsx
@@ -1,11 +1,10 @@
-import React from 'react'
 import Paper from '@material-ui/core/Paper';
 import { makeStyles } from '@material-ui/core/styles';
-import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
-import PhoneIcon from '@material-ui/icons/Phone';
+import Tabs from '@material-ui/core/Tabs';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 import PersonPinIcon from '@material-ui/icons/PersonPin';
+import React from 'react';
 
 const useStyles = makeStyles({
   root: {
@@ -13,7 +12,6 @@ const useStyles = makeStyles({
     // maxWidth: 500,
   },
 });
-
 
 const ProfileTabs = ({handleChange,query}:any) => {
   const classes = useStyles();
