--- frontend/src/components/test/LoginButton.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { useAuth0 } from "@auth0/auth0-react";
-
-const LoginButton = () => {
-  const {
-    isAuthenticated,
-    loginWithRedirect,
-  } = useAuth0();
-
-  return  (
-    <button onClick={loginWithRedirect}>Log in</button>
-  );
-};
-
-export default LoginButton;
