--- frontend/src/components/test/LogoutButton.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import { useAuth0 } from "@auth0/auth0-react";
-
-const LogoutButton = () => {
-  const { logout } = useAuth0();
-
-  return (
-    <button onClick={() => logout({ returnTo: window.location.origin })}>
-      Log Out します
-    </button>
-  );
-};
-
-export default LogoutButton;
