+++ frontend/src/hooks/castomHook/useMutatePost.ts
@@ -1,14 +1,11 @@
 import { EditPost, PostData } from '../../types/types'
 import axios from 'axios'
 import {useAppSelector, useAppDispatch } from '../../app/hooks'
-// import { resetEditedTask } from '../slices/todoSlice'
 import { useQueryClient, useMutation } from 'react-query'
 import { selectSpot } from '../../slices/spotSlice'
-import { setEditedPost,resetEditedPost, setShowPost } from '../../slices/postSlice'
-
 import { useMutateSpot } from './useMutateSpot'
-
 import { selectHeaders } from "../../slices/headersSlice";
+import { resetEditedPost } from '../../slices/postSlice'
 
 export const useMutatePost = () => {
   const dispatch = useAppDispatch()
@@ -32,7 +29,7 @@ export const useMutatePost = () => {
             res.data,
           ])
         }
-        // dispatch(resetEditedPost())
+        dispatch(resetEditedPost())
       },
     }
   )
