+++ frontend/src/hooks/castomHook/useMutateUser.ts
@@ -1,16 +1,8 @@
-import { EditPost, PostData } from '../../types/types'
 import axios from 'axios'
 import {useAppSelector, useAppDispatch } from '../../app/hooks'
-// import { resetEditedTask } from '../slices/todoSlice'
-import { useQueryClient, useMutation } from 'react-query'
-import { selectSpot } from '../../slices/spotSlice'
-import { setEditedPost,resetEditedPost, setShowPost } from '../../slices/postSlice'
-
-import { useMutateSpot } from './useMutateSpot'
-
+import { useMutation } from 'react-query'
 import { selectHeaders } from "../../slices/headersSlice";
 import { setShowProfile } from '../../slices/profileSlice'
-
 
 export const useMutateUser = () => {
   const dispatch = useAppDispatch()
@@ -36,14 +28,7 @@ export const useMutateUser = () => {
   )
   const profileUpdateMutation = useMutation(
     (data:any) =>
-    axios.put(`${process.env.REACT_APP_REST_URL}/profiles/2`,data,headers),
-    {
-      onSuccess: (res) => {
-        console.log(res)
-      }      
-    }
-    
-    
+    axios.put(`${process.env.REACT_APP_REST_URL}/profiles/1`,data,headers),
     )
 
     return { userIdMutation,profileUpdateMutation,profileShowMutation}
