+++ frontend/src/hooks/reactQuery/useQueryPostShow.ts
@@ -1,20 +1,11 @@
-import React, { useEffect, useState } from "react";
-import { useQuery } from 'react-query'
-import axios from 'axios'
-import { PostData, User } from '../../types/types'
-import { useAuth0 } from "@auth0/auth0-react";
-import { useAppSelector } from "../../app/hooks";
-import { Auth0Provider } from '@auth0/auth0-react';
-import { selectHeaders } from "../../slices/headersSlice";
+import axios from 'axios';
+import { useQuery } from 'react-query';
 
 export const useQueryPostShow = (id:number) => {
-  // const token = useAppSelector(selectUserToken)
-
-  const headers = useAppSelector(selectHeaders)
 
   const getPostShow = async () => {
     const { data } = await axios.get(
-      `${process.env.REACT_APP_REST_URL}/posts/${id}`,headers)
+      `${process.env.REACT_APP_REST_URL}/posts/${id}`)
       return data
   }
 
@@ -23,12 +14,9 @@ export const useQueryPostShow = (id:number) => {
     queryFn: getPostShow,
     staleTime: 0,
     cacheTime: 0,
-    // refetchOnWindowFocus: true,
   })
 }
 
-// postのshowと同時にspotも必要
 
-// 一緒に持ってくる?
