+++ frontend/src/hooks/reactQuery/useQueryPosts.ts
@@ -1,17 +1,7 @@
-import React, { useEffect, useState } from "react";
-import { useQuery } from 'react-query'
-import axios from 'axios'
-import { PostData, User } from '../../types/types'
-import { useAuth0 } from "@auth0/auth0-react";
-import { useAppSelector } from "../../app/hooks";
-import { Auth0Provider } from '@auth0/auth0-react';
-import { selectHeaders } from "../../slices/headersSlice";
+import axios from 'axios';
+import { useQuery } from 'react-query';
 
 export const useQueryPosts = () => {
-  // const token = useAppSelector(selectUserToken)
-
-  const headers = useAppSelector(selectHeaders)
-
   const getPosts = async () => {
 
     const { data } = await axios.get(
@@ -25,8 +15,5 @@ export const useQueryPosts = () => {
     queryKey: 'posts',
     queryFn: getPosts,
     staleTime: 30000,
-    // refetchOnWindowFocus: true,
-    //cacheTime: 5000,
-    // refetchInterval: 5000,
   })
 }
