+++ frontend/src/hooks/reactQuery/useQueryProfileShow.ts
@@ -1,15 +1,10 @@
-import React, { useEffect, useState } from "react";
-import { useQuery } from 'react-query'
-import axios from 'axios'
-import { PostData, User } from '../../types/types'
-import { useAuth0 } from "@auth0/auth0-react";
-import { useAppSelector,useAppDispatch } from "../../app/hooks";
-import { Auth0Provider } from '@auth0/auth0-react';
+import axios from 'axios';
+import { useQuery } from 'react-query';
+import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import { selectHeaders } from "../../slices/headersSlice";
 import { setEditedProfile } from "../../slices/profileSlice";
 
 export const useQueryProfileShow = (id:number) => {
-  // const token = useAppSelector(selectUserToken)
 
   const headers = useAppSelector(selectHeaders)
   const dispatch = useAppDispatch()
@@ -26,12 +21,7 @@ export const useQueryProfileShow = (id:number) => {
     queryFn: getProfileShow,
     staleTime: 0,
     cacheTime: 0,
-    // refetchOnWindowFocus: true,
   })
 }
-
-// postのshowと同時にspotも必要
-
-// 一緒に持ってくる?
