+++ frontend/src/index.tsx
@@ -1,12 +1,11 @@
+import { Auth0Provider } from '@auth0/auth0-react';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import { Provider } from 'react-redux';
 import App from './App';
 import { store } from './app/store';
-import { Provider } from 'react-redux';
+import './index.css';
 import * as serviceWorker from './serviceWorker';
-
-import { Auth0Provider } from '@auth0/auth0-react';
 
 const domain:any = process.env.REACT_APP_AUTH0_DOMAIN;
 const clientId:any = process.env.REACT_APP_AUTH0_CLIENT_ID;
@@ -18,9 +17,8 @@ ReactDOM.render(
   clientId={clientId}
   redirectUri={window.location.origin}
   audience={audience}
-  >
-    
-    <Provider store={store}>
+  > 
+  <Provider store={store}>
       <App />
     </Provider>
   </Auth0Provider>,
