+++ frontend/src/slices/headersSlice.ts
@@ -1,15 +1,9 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { RootState } from '../app/store';
-// import { EditTask, Tag } from '../types/types'
 
 export interface TokenState {
   token: string
 }
-
-// export interface TaskState {
-//   editedTask: EditTask
-//   editedTag: Tag
-// }
 
 const initialState:any = {  
   headers: {
@@ -17,12 +11,7 @@ const initialState:any = {
     'Content-Type': 'application/json',
     // "sub": 'google-oauth2|106160814069089305764'
   }
-  }
-
-// export const userTokenSlice = createSlice({
-//   name:'userToken',
-
-// })
+}
 
 export const headersSlice = createSlice({
   name: 'headers',
