+++ frontend/tailwind.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   darkMode: false,
+
   theme: {
       extend: {},
   },
